// src/jupyter/client/generated/models/ContentsSaveRequest.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * Jupyter Server API
 * Server API
 *
 * The version of the OpenAPI document: 5
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ContentsSaveRequest
 */
export interface ContentsSaveRequest {
    /**
     * The new filename if changed
     * @type {string}
     * @memberof ContentsSaveRequest
     */
    name?: string;
    /**
     * New path for file or directory
     * @type {string}
     * @memberof ContentsSaveRequest
     */
    path?: string;
    /**
     * Path dtype ('notebook', 'file', 'directory')
     * @type {string}
     * @memberof ContentsSaveRequest
     */
    type?: string;
    /**
     * File format ('json', 'text', 'base64')
     * @type {string}
     * @memberof ContentsSaveRequest
     */
    format?: string;
    /**
     * The actual body of the document excluding directory type
     * @type {string}
     * @memberof ContentsSaveRequest
     */
    content?: string;
}

/**
 * Check if a given object implements the ContentsSaveRequest interface.
 */
export function instanceOfContentsSaveRequest(value: object): value is ContentsSaveRequest {
    return true;
}

export function ContentsSaveRequestFromJSON(json: any): ContentsSaveRequest {
    return ContentsSaveRequestFromJSONTyped(json, false);
}

export function ContentsSaveRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ContentsSaveRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'path': json['path'] == null ? undefined : json['path'],
        'type': json['type'] == null ? undefined : json['type'],
        'format': json['format'] == null ? undefined : json['format'],
        'content': json['content'] == null ? undefined : json['content'],
    };
}

export function ContentsSaveRequestToJSON(json: any): ContentsSaveRequest {
    return ContentsSaveRequestToJSONTyped(json, false);
}

export function ContentsSaveRequestToJSONTyped(value?: ContentsSaveRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'path': value['path'],
        'type': value['type'],
        'format': value['format'],
        'content': value['content'],
    };
}